// src/app/model/indent-model.ts
export class IndentModal {
    ordered_for: any;
    ordered_by: any;
    on_behalf_of: any;
    category: any;
    products: any;
    ordered_forother: any;
    units: any;
    quantity: any;
    price: any;
    expected_date: any;
    delivered_date: any;
    place: any;
    pincode: any;
    district: any;
    state: any;
    total_price: any;
    total_order: any;
}